DV-24: Put data stratification into TreemapView.tsx, it is more applicable there.
