+++ src/containers/App.tsx
@@ -1,10 +1,7 @@
 import 'ihme-ui/dist/ihme-ui.css';
 
 import { json } from 'd3-fetch';
-import {
-  HierarchyNode,
-  stratify,
-} from 'd3-hierarchy';
+
 import {
   LoadingIndicator,
   ResponsiveContainer,
@@ -15,7 +12,7 @@ import DropShadowDefs from '../components/DropShadowDefs';
 import ResponsiveSVG from '../components/ResponsiveSVG';
 import TreemapView from './TreemapView';
 
-interface RawDatum {
+export interface RawDatum {
   location_name: string;
   location_id: number;
   parent_location_id: number;
@@ -25,7 +22,7 @@ interface RawDatum {
 }
 
 interface AppState {
-  data: HierarchyNode<RawDatum>;
+  data: RawDatum[];
 }
 
 export default class App
@@ -33,18 +30,6 @@ extends React.PureComponent<
   {},
   AppState
 > {
-  static stratifyData(data: RawDatum[]): HierarchyNode<RawDatum> {
-    return stratify<RawDatum>()
-      .id((d: RawDatum) => String(d.location_id))
-      .parentId((d: RawDatum) => (
-        d.parent_location_id
-        ? String(d.parent_location_id)
-        : null
-      ))(data)
-      .sum(d => d.value)
-      .sort((a, b) =>  (b.height - a.height || b.value - a.value));
-  }
-
   constructor(props) {
     super(props);
     this.state = { data: null };
@@ -52,7 +37,7 @@ extends React.PureComponent<
 
   async componentDidMount() {
     const data = await json('resources/mock_populations_with_attributes.json');
-    this.setState({ data: App.stratifyData(data as RawDatum[]) });
+    this.setState({ data: data as RawDatum[] });
   }
 
   renderView(data) {
